LKF foundations
We provide a very explicit and transparent (but method-less, and
therefore fairly safe) definition of LKF formulas. Interestingly,
quantifiers are represented by variants with first-order functions,
replacing e.g. the bindings of lambda Prolog. Another common theme
is that dependencies between modules are currently explicit and
anchored to one particular module, whereas ideally we will want to
abstract these details and make them generic.

In addition to formulas, contexts and sequents for the LKF sequent
calculus are provided as well. Contexts in particular benefit from
a modest amount of encapsulation that hides the (only slightly)
more complicated implementation details and exposes a homogeneous
interface.