Certificate framework
Here we introduce the missing components of the framework and pave
the way for the functional implementation of the kernel. First we
defined two "generic" components: the braindead choice type (which
is so simple it may well be defined elsewhere, if not for a
dogmatic approach to type declaration and separation), and the
certificate type. Again, in the interest of agility, we provide one
particular instance of FPC in the form of certificate constructors:
that of the "canonical form" that is expected to reproduce the
proof tree exactly.

The "bureau" (i.e. collection of clerks and experts) that
implements an FPC for the LKF system is also defined. An
implementation for the explicit proof tree certificate constructors
described above is also trivially provided. This will be one of the
main points of abstraction later on. No particular effort has been
made to make patterns exhaustive, and catch-alls here have been
considered acceptable and not affecting soundness. The mapping is
predictably straightforward.