LKF kernel
Assembling the pieces of the puzzle and sidestepping complex
dependencies, the final piece of the basic checker is added. The
kernel offers a recursive checking function, currently cut-free,
that checks and enforces some of the requirements of a functional
interpretation. The semantics of the model enforces much of this,
but certain checks are still needed.

For each case, a helper function is defined to encode the
corresponding inference rule of the sequent calculus and perform
the recursive call to the main function. Each of the helpers relies
on a call to the corresponding clerk or expert in the provided
bureau, analyzing and propagating the results according to the
inference rule.

This completes the basic implementation of the checker.