Canonical certificates
The following are provided to define canonical certificates:

 - Certificate definitions.

 - Index constructors.

 - Clerks and experts (the bureau).

In addition, atoms are needed to compose formulas.

A dummy implementation of terms is also provided, but not used,
with atoms being fully propositional. Terms need to be stubbed,
however.

As test we give a canonical certificate obtained from a proof by
resolution.

Furthermore, the universe