Add new automated resolution elaboration
A slightly more interesting example from the test suite of
the Checkers tool is added.  Naturally, atoms and terms must be
extended with the corresponding risk of name clashes and necessity
of renaming.  This illustrates the limits of the approach and the
need for a more systematic build discipline.  Nonetheless, this
is indirect proof of automatic generation of test cases in an
early stage, the automation of which is underway.

Naturally, the same issues arise for the definition of the
certificates, indexes and so on. To what extent these can proceed
from the lambda Prolog source to the necessary OCaml files, with
their various naming conventions, typing disciplines, modular
constructs, etc., will determine the nature of the final process.